farmchain/ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
Pillow==10.3.0
numpy==1.26.4
scikit-learn==1.4.2
pydantic==2.7.1
httpx==0.27.0
python-dotenv==1.0.1
